+++ src/components/index.js
@@ -8,5 +8,6 @@ export { default as InstancePlugin } from './InstancePlugin';
 export { default as ConnectedTitle } from './ConnectedTitle';
 export { default as UnconnectedTitle } from './UnconnectedTitle';
 export { default as TitleLabel } from './TitleLabel';
+export { default as TitleField } from './TitleField';
 export { default as TitlesView } from './TitlesView';
 export { PaneLoading, ViewLoading } from './Loading';