+++ src/edit/succeedingTitleFields.js
@@ -1,45 +1,32 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
+import { FieldArray } from 'react-final-form-arrays';
 
-import { TextField } from '@folio/stripes/components';
+import { RepeatableField } from '@folio/stripes/components';
 
-import RepeatableField from '../components/RepeatableField';
+import { TitleField } from '../components';
 
-const SucceedingTitles = props => {
-  const {
-    canAdd,
-    canEdit,
-    canDelete,
-  } = props;
-
-  return (
-    <RepeatableField
-      name="succeedingTitles"
-      label={<FormattedMessage id="ui-inventory.succeedingTitles" />}
-      addLabel={<FormattedMessage id="ui-inventory.addSucceedingTitle" />}
-      addButtonId="clickable-add-succeedingTitle"
-      template={[
-        {
-          name: 'subInstanceId',
-          label: 'FOLIO ID',
-          component: TextField,
-          required: true,
-          disabled: !canEdit
-        },
-      ]}
-      newItemTemplate={{ subInstanceId: '' }}
-      canAdd={canAdd}
-      canDelete={canDelete}
-    />
-  );
-};
+const SucceedingTitles = ({ canAdd, canEdit, canDelete }) => (
+  <FieldArray
+    addLabel={<FormattedMessage id="ui-inventory.addSucceedingTitle" />}
+    legend={<FormattedMessage id="ui-inventory.succeedingTitles" />}
+    id="clickable-add-succeedingTitle"
+    component={RepeatableField}
+    name="succeedingTitles"
+    canAdd={canAdd}
+    canRemove={canDelete}
+    canEdit={canEdit}
+    renderField={(field, index, fields) => <TitleField field={field} index={index} fields={fields} titleIdKey="succeedingInstanceId" />}
+  />
+);
 
 SucceedingTitles.propTypes = {
   canAdd: PropTypes.bool,
   canEdit: PropTypes.bool,
   canDelete: PropTypes.bool,
 };
+
 SucceedingTitles.defaultProps = {
   canAdd: true,
   canEdit: true,